total refactoring of IQ blobs and sweeps
