+++ qualibration_graphs/quantum_dots/calibration_utils/iq_utils/iq_sweep/__init__.py
@@ -1,5 +1,5 @@
-from .parameters import Parameters
 from .analysis import (
+    FitParameters,
     process_raw_dataset,
     fit_raw_data,
     fit_raw_data_pca_gaussian,
@@ -11,10 +11,13 @@ from .plotting import (
     plot_sweep_summary,
     plot_metric_vs_sweep,
     plot_histograms_vs_sweep,
+    plot_rotated_iq_density,
+    plot_rotated_iq_density_at_optimum,
+    plot_single_histogram_with_fit,
 )
 
 __all__ = [
-    "Parameters",
+    "FitParameters",
     "process_raw_dataset",
     "fit_raw_data",
     "fit_raw_data_pca_gaussian",
@@ -24,4 +27,7 @@ __all__ = [
     "plot_sweep_summary",
     "plot_metric_vs_sweep",
     "plot_histograms_vs_sweep",
+    "plot_rotated_iq_density",
+    "plot_rotated_iq_density_at_optimum",
+    "plot_single_histogram_with_fit",
 ]
