+++ qualibration_graphs/quantum_dots/calibration_utils/iq_utils/parameters.py
@@ -1,14 +1,15 @@
 from typing import Literal
-from qualibrate import NodeParameters
 from qualibrate.core.parameters import RunnableParameters
-from qualibration_libs.parameters import CommonNodeParameters, QubitsExperimentNodeParameters, QubitPairExperimentNodeParameters
 
+__all__ = ["IQParameters", "IQSweepParameters"]
 
-class NodeSpecificParameters(RunnableParameters):
-    num_shots: int = 2000
+class IQParameters(RunnableParameters): 
+    um_shots: int = 2000
     """Number of runs to perform per sweep point. Default is 2000."""
     operation: Literal["readout", "readout_QND"] = "readout"
     """Type of operation to perform. Default is "readout"."""
+
+class IQSweepParameters(IQParameters): 
     sweep_name: str = "detuning"
     """Name of the swept coordinate in ds_raw (e.g. "detuning", "integration_time")."""
     optimization_metric: Literal["fidelity", "visibility"] = "fidelity"
@@ -19,13 +20,3 @@ class NodeSpecificParameters(RunnableParameters):
     Rabi-style IQ-blob experiment) or a single mixed-state acquisition (I,Q,
     as in a PSB search where loading is random). Determines whether a confusion
     matrix is computed. Default False = mixed-state mode."""
-
-
-class Parameters(
-    NodeParameters,
-    CommonNodeParameters,
-    NodeSpecificParameters,
-    QubitsExperimentNodeParameters,
-    QubitPairExperimentNodeParameters,
-):
-    pass
