--- tests/.qm_cluster_config.json.example
@@ -1,4 +1,0 @@
-{
-  "host": "YOUR_CLUSTER_IP",
-  "cluster_name": "YOUR_CLUSTER_NAME"
-}
